# api-gestao-propostas/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestão de Propostas",
        "description": "API REST para gestão de clientes e propostas. Suporta idempotência via header Idempotency-Key e optimistic lock em atualizações.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local"
        },
        {
            "url": "/api/v1",
            "description": "Relativo"
        }
    ],
    "paths": {
...
        "/propostas": {
            "get": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Listar propostas",
                "description": "Lista propostas com filtros, ordenação e paginação",
                "operationId": "2a0f96c5924c255560681a447cdb80b3",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "DRAFT",
                                "SUBMITTED",
                                "APPROVED",
                                "REJECTED",
                                "CANCELED"
                            ],
                            "example": "DRAFT"
                        }
                    },
                    {
                        "name": "cliente_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "produto",
                        "in": "query",
                        "description": "Busca parcial (ILIKE)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Seguro"
                        }
                    },
                    {
                        "name": "valor_min",
                        "in": "query",
                        "description": "Valor mensal mínimo",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 100
                        }
                    },
                    {
                        "name": "valor_max",
                        "in": "query",
                        "description": "Valor mensal máximo",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "example": 500
                        }
                    },
                    {
                        "name": "ordenar_por",
                        "in": "query",
                        "description": "Campo para ordenação",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "created_at",
                                "valor_mensal",
                                "produto",
                                "status"
                            ],
                            "example": "created_at"
                        }
                    },
                    {
                        "name": "direcao",
                        "in": "query",
                        "description": "Direção da ordenação",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ],
                            "example": "desc"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página (máx. 100, default 15)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de propostas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Criar proposta",
                "operationId": "c89c6b9895eb084a1ec21a264be20669",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "produto",
                                    "valor_mensal"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "produto": {
                                        "type": "string",
                                        "example": "Seguro Auto"
                                    },
                                    "valor_mensal": {
                                        "type": "number",
                                        "example": 150
                                    },
                                    "origem": {
                                        "description": "Opcional, default: API",
                                        "type": "string",
                                        "enum": [
                                            "APP",
                                            "SITE",
                                            "API"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proposta criada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação ou regra de negócio",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/propostas/{proposta}": {
            "get": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Obter proposta por ID",
                "operationId": "7b3c3a10d792350035c57064f663682e",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "description": "ID da proposta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposta encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proposta não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Atualizar proposta",
                "description": "Atualiza campos da proposta em DRAFT. Requer versão para optimistic lock. Retorna 409 se versão desatualizada.",
                "operationId": "f56d5024340d7d19bd85cc094d2b98e7",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "versao"
                                ],
                                "properties": {
                                    "versao": {
                                        "description": "Versão atual para optimistic lock",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "produto": {
                                        "type": "string",
                                        "example": "Seguro Residencial"
                                    },
                                    "valor_mensal": {
                                        "type": "number",
                                        "example": 199.9
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proposta atualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflito de versão",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Apenas propostas em DRAFT podem ser editadas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/propostas/{proposta}/submit": {
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Submeter proposta",
                "description": "Transiciona DRAFT → SUBMITTED. Aceita Idempotency-Key para evitar duplicatas em retries.",
                "operationId": "a9eb9d508f96c4534357576f5208d6cb",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "Idempotency-Key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposta submetida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Transição inválida ou proposta não está em DRAFT",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/propostas/{proposta}/approve": {
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Aprovar proposta",
                "description": "Transiciona SUBMITTED → APPROVED",
                "operationId": "b909cdeaec6ccff41d240855d0780027",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposta aprovada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Transição inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/propostas/{proposta}/reject": {
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Rejeitar proposta",
                "description": "Transiciona SUBMITTED → REJECTED",
                "operationId": "c95058f26ae226e48c79df763bab678f",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposta rejeitada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Transição inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/propostas/{proposta}/cancel": {
            "post": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Cancelar proposta",
                "description": "Transiciona SUBMITTED → CANCELED",
                "operationId": "ae9342932a814dcaaaf17b4a3baa0d8e",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Proposta cancelada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Proposta"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Transição inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/propostas/{proposta}/auditoria": {
            "get": {
                "tags": [
                    "Propostas"
                ],
                "summary": "Histórico de auditoria",
                "description": "Retorna o histórico de alterações da proposta",
                "operationId": "844eabf98001d91e73d9910544e3c872",
                "parameters": [
                    {
                        "name": "proposta",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de eventos de auditoria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PropostaAuditoria"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proposta não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Proposta": {
                "title": "Proposta",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "cliente_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "produto": {
                        "type": "string",
                        "example": "Seguro Auto"
                    },
                    "valor_mensal": {
                        "type": "number",
                        "format": "float",
                        "example": 150
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "DRAFT",
                            "SUBMITTED",
                            "APPROVED",
                            "REJECTED",
                            "CANCELED"
                        ]
                    },
                    "origem": {
                        "type": "string",
                        "enum": [
                            "APP",
                            "SITE",
                            "API"
                        ]
                    },
                    "versao": {
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "title": "Error",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Mensagem de erro"
                    },
                    "errors": {
                        "description": "Detalhes de validação (quando aplicável)",
                        "type": "object"
                    }
                },
                "type": "object"
            }
        }
    },
...
}